requirements.txt
ansicon==1.89.0
APScheduler==3.6.3
asgiref
blessed
certifi==2020.4.5.1
cffi
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
cryptography
pyparsing
xlsxwriter
pyPdf
openpyxl

Django
django-appconf
django-base-url==2019.7.24
django-cors-headers
django-crispy-forms
django-cryptography
django-uuidfield==0.5.0
django-xhtml2pdf==0.0.4
html5lib==1.0.1
idna==2.9
jinxed==1.0.0
numpy
pdfrw==0.4
Pillow
pycparser
PyPDF2==1.26.0
zeep
pytesseract==0.3.4
pytz
reportlab
requests==2.23.0
six
sqlparse
typesentry==0.2.7
tzlocal
urllib3==1.25.9
virtualenv
virtualenvwrapper-win
wcwidth
webencodings==0.5.1
xhtml2pdf==0.2.4
opencv-python
gunicorn
